refactor(server): move the shutdown save sequence onto Server
was a private fn in the binary, so embedders could only copy it, and a copy
stops saving whatever steel later adds to the sequence.

body moves verbatim apart from the receiver and two comments that restated the
log line below them. doc records the packet-processing precondition.
