+++ steel-core/src/server/run_loop.rs
@@ -4,11 +4,13 @@ use super::{
     COMMAND_REQUESTS_PER_TICK, COMMAND_RESUMPTIONS_PER_TICK, CancellationToken, ChunkPos,
     ChunkSender, CommandExecutionContext, CommandExecutionOwner, CommandRequest,
     CommandResultCallback, CommandSender, CommandSource, Duration, EncodedChunk,
-    ExecutionCommandSource, ExecutionStop, GameTickTaskGuard, Instant, JoinSet, NetworkConnection,
-    PendingCommandExecutionQueue, Player, SEND_PLAYER_INFO_INTERVAL, SLOW_CHUNK_TICK_THRESHOLD,
-    Server, StringReader, SuggestionError, Suggestions, TAB_LIST_UPDATE_INTERVAL, TabListTickStats,
-    ThreadPool, World, command_suggestions_packet, sleep, spawn_blocking,
+    ExecutionCommandSource, ExecutionStop, GameTickTaskGuard, GlobalPlayerData, Instant, JoinSet,
+    MenuRemovalStatus, NetworkConnection, PendingCommandExecutionQueue, PersistentPlayerData,
+    Player, SEND_PLAYER_INFO_INTERVAL, SLOW_CHUNK_TICK_THRESHOLD, Server, StringReader,
+    SuggestionError, Suggestions, TAB_LIST_UPDATE_INTERVAL, TabListTickStats, ThreadPool, World,
+    command_suggestions_packet, sleep, spawn_blocking,
 };
+use steel_registry::vanilla_custom_stats;
 use steel_utils::threading::{available_worker_threads, worker_threads_for_available};
 
 impl Server {
@@ -73,6 +75,91 @@ impl Server {
                 log::error!("{task} task failed: {error}");
             }
         }
+    }
+
+    /// Saves everything and tears the worlds down, in the order shutdown requires.
+    ///
+    /// # Panics
+    ///
+    /// Panics if a player's menus cannot be removed, which means packets are still being
+    /// processed. Callers must stop packet processing before calling this; the standalone
+    /// server does so by closing its task tracker and awaiting the drain.
+    pub async fn save_and_shutdown(&self) {
+        if let Err(error) = self.flush_known_players().await {
+            log::error!("Failed to flush known player cache during shutdown: {error}");
+        }
+
+        let players = self.get_players();
+        for player in &players {
+            player.close_connection();
+            assert_eq!(
+                player.remove_all_menus(),
+                MenuRemovalStatus::Complete,
+                "shutdown menu removal must run after packet processing stops"
+            );
+        }
+
+        for world in self.worlds.values() {
+            world.chunk_map.stop_generation_refill_loop();
+            world.chunk_map.task_tracker.close();
+            world.chunk_map.task_tracker.wait().await;
+        }
+
+        let mut players_to_save = Vec::new();
+        for player in players {
+            let domain = player.get_world().domain().to_owned();
+            player.award_custom_stat(&vanilla_custom_stats::LEAVE_GAME);
+            let data = PersistentPlayerData::from_player(&player);
+            player.store_ender_pearls_with_player();
+            players_to_save.push((player, domain, data));
+        }
+
+        log::info!("Saving world data...");
+        let command_data = self.save_command_data().await;
+        match command_data.scoreboards {
+            Ok(saved) => log::info!("Saved {saved} domain scoreboards"),
+            Err(error) => log::error!("Failed to save domain scoreboards: {error}"),
+        }
+        match command_data.storage {
+            Ok(saved) => log::info!("Saved {saved} domain command storages"),
+            Err(error) => log::error!("Failed to save domain command storage: {error}"),
+        }
+        let mut total_saved = 0;
+        for world in self.worlds.values() {
+            world.cleanup(&mut total_saved).await;
+        }
+        log::info!("Saved {total_saved} chunks");
+
+        log::info!("Saving player data...");
+        let mut saved = 0;
+        for (player, domain, data) in players_to_save {
+            let uuid = player.gameprofile.id;
+            match self
+                .player_data_storage
+                .save_domain_data(&domain, uuid, &data)
+                .await
+            {
+                Ok(()) => {
+                    saved += 1;
+                }
+                Err(e) => {
+                    log::error!("Failed to save player {uuid} domain data during shutdown: {e}");
+                }
+            }
+            if let Err(e) = self
+                .player_data_storage
+                .save_global(
+                    uuid,
+                    &GlobalPlayerData {
+                        last_active_domain: domain,
+                    },
+                )
+                .await
+            {
+                log::error!("Failed to save player {uuid} global data during shutdown: {e}");
+            }
+        }
+        log::info!("Saved {saved} players");
     }
 
     /// The main game tick loop (20 TPS, governed by tick rate manager).
